backend/requirements.txt
azure-cognitiveservices-speech==1.7.0
